requirements.txt
selenium==4.16.0
openai==1.12.0
pandas==2.2.0
numpy==1.26.3
flask==3.0.1
requests==2.31.0
python-dotenv==1.0.0
schedule==1.2.1
discord.py==2.3.2
pytest==7.4.4
black==24.1.1
pylint==3.0.3 